+++ js/data.js
@@ -1,0 +1,62 @@
+import {getRandomInRange, getRandomArrayElement} from "./util.js";
+
+const PHOTOGRAPHIES_COUNT = 25
+const COMMENTS_COUNT = 3
+
+const NAMES = [
+  'Артём',
+  'Виктор',
+  'Мария',
+  'Никита',
+  'Ксения',
+  'Илья',
+  'Андрей',
+  'Апполонария',
+  'Фелиция',
+  'Канеки',
+]
+
+const DESCRIPTIONS = [
+  'Крутышка!',
+  'Не круто',
+  'Крутяк',
+  'Отстой',
+  'норм бравлик',
+  'бебрики',
+  'фуу',
+  'ку челики чд кд',
+]
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+]
+
+let commentId = 0
+const createComment = () => {
+  return {
+    id: commentId++,
+    avatar: `img/avatar-${getRandomInRange(1, 6)}.svg`,
+    message: getRandomArrayElement(MESSAGES),
+    name: getRandomArrayElement(NAMES),
+  }
+}
+
+let currentId = 0
+const createPhotography = () => {
+  return {
+    id: ++currentId,
+    url: `photos/${currentId}.jpg`,
+    description: getRandomArrayElement(DESCRIPTIONS),
+    likes: getRandomInRange(15, 200),
+    comments: Array.from({length: COMMENTS_COUNT}, createComment),
+  }
+}
+
+const createPhotographies = () => Array.from({length: PHOTOGRAPHIES_COUNT}, createPhotography)
+
+export {createPhotographies};
